# docs/api_doc.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/surveys/templates/{template_id}": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Get Survey Template",
        "description": "- Open to survey operators.\n- Returns template draft detail and latest published version reference.",
        "operationId": "get_survey_template_api_v1_surveys_templates__template_id__get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyTemplateDetailQM"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Surveys"
        ],
        "summary": "Update Survey Template",
        "description": "- Open to survey operators.\n- Updates template draft data without mutating published versions.",
        "operationId": "update_survey_template_api_v1_surveys_templates__template_id__patch",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSurveyTemplateRequestPydantic"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionType": {
        "type": "string",
        "enum": [
          "single_choice",
          "multi_choice",
          "text"
        ],
        "title": "QuestionType"
      },
...
      "SimpleErrorResponseModel": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "SimpleErrorResponseModel"
      },
...
      "SurveyTemplateDetailQM": {
        "properties": {
          "id_": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/SurveyTemplateQuestionQM"
            },
            "type": "array",
            "title": "Questions"
          },
          "latest_published_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Version Id"
          }
        },
        "type": "object",
        "required": [
          "id_",
          "name",
          "questions",
          "latest_published_version_id"
        ],
        "title": "SurveyTemplateDetailQM"
      },
...
      "SurveyTemplateQuestionQM": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "question_type": {
            "type": "string",
            "title": "Question Type"
          },
          "required": {
            "type": "boolean",
            "title": "Required"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "key",
          "title",
          "question_type",
          "required",
          "options"
        ],
        "title": "SurveyTemplateQuestionQM"
      },
      "UpdateSurveyTemplateRequestPydantic": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/app__presentation__http__controllers__surveys__create_template__SurveyQuestionPydantic"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "name",
          "questions"
        ],
        "title": "UpdateSurveyTemplateRequestPydantic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "app__presentation__http__controllers__surveys__create_template__SurveyQuestionPydantic": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "required": {
            "type": "boolean",
            "title": "Required"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "key",
          "title",
          "question_type",
          "required"
        ],
        "title": "SurveyQuestionPydantic"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  }
}